Adds message URI conversion
Converts a generic message uri that a client wants to dereference into a
local one on the default WoN node.

Optionally, the client can send a URL parameter `wonnode=[wonNodeUri]`
to ask for conversion on that WoN node. If that fails, the default WoN
node is used for conversion.

If the conversion fails, the resourceUri is left as-is.